% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlogitnorm.R
\name{dlogitnorm}
\alias{dlogitnorm}
\title{Logit-normal distribution functions}
\usage{
dlogitnorm(x, location = 0, scale = 1)
}
\arguments{
\item{x}{input in [0,1]}

\item{location}{is the location parameter}

\item{scale}{is the scale parameter}

\item{n}{number of random variables to generate}
}
\value{
Returns a list of results from analysis.
}
\description{
The logit-normal distribution is a probability distribution of a random variable
whose logit has a normal distribution. If Y is a random variable with a normal
distribution, and P is the logistic function, then X = P(Y) has a logit-normal
distribution; likewise, if X is logit-normally distributed, then
Y = logit(X)= log (X/(1-X)) is normally distributed.
}
\examples{
x=seq(0,1,0.1)
y=dlogitnorm(x)
plot(x,y)


}
\author{
Daniel D Sjoberg \email{sjobergd@mskcc.org}
}
