% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.functions.R
\name{erf}
\alias{erf}
\title{Misc. Functions
Error Function and Complimentary Error Function
Logit and Inverse Logit Function}
\usage{
erf(x)
}
\arguments{
\item{x}{function input}
}
\value{
function output
}
\description{
Misc. Functions
Error Function and Complimentary Error Function
Logit and Inverse Logit Function
}
\examples{
erf(0)
}
\author{
Daniel D Sjoberg \email{sjobergd@mskcc.org}
}
