% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_search.R
\name{golden_search}
\alias{golden_search}
\title{One-diminsional Search Algorithm}
\usage{
golden_search(x, a, b, tol = 1e-04)
}
\arguments{
\item{x}{function with one argument}

\item{a}{min value of search values}

\item{b}{max value of search values}

\item{tol}{tolerance level. Default is \code{1e-04}}
}
\description{
Find the minimum of a function using the Golded-section search algorithm
}
\examples{
contfunc = function(x){
  -exp(-x) * sin(x)
}
golden_search(contfunc, a = 0, b = 1.5)
}
