% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjofun-package.R
\docType{package}
\name{sjofun-package}
\alias{sjofun}
\alias{sjofun-package}
\title{sjofun: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
