% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.diff.R
\name{model.diff}
\alias{model.diff}
\title{Model-based test for treatment efficacy}
\usage{
model.diff(data, outcome, covars, model, cov, coef, type = "logistic",
  output.details = FALSE)
}
\arguments{
\item{data}{a data frame containing the outcome of the outcome predictions.}

\item{outcome}{the outcome, or response variable name. Must be a variable contained within the data frame specified in data=.}

\item{covars}{vector of covariate/predictor variable(s) names. Must be a variable(s) contained within the data frame specified in data=.  If model includes an intercept, user must include the column of ones in the covars vector}

\item{model}{glm object of the predictive model estimated on historical cohort.  If specified, the outcome, covars, cov, and coef objects will be extracted from object.}

\item{cov}{Variance-covaraince matrix of the beta coefficients from predictive model.}

\item{coef}{Vector of the beta coefficients from predictive model.  If model includes an intercept, a vector of ones must appear in data.}

\item{type}{Type of predictive model used.  logistic is currently the only valid input.}

\item{output.details}{Save additional information.  Default is FALSE.}
}
\value{
Returns a list of results from analysis.
}
\description{
The model-based sinlge-arm comparison addresses the fundamental question of whether the current treatment provides a clinically significant 
improvement over prior treatments in the population. The proposed test statistic computes the difference between the 
observed outcome from the current treatment and the covariate-specific predicted outcome based on a model of the 
historical data. Thus, the difference between the observed and predicted quantities is attributed to the current 
treatment.
}
\details{
Heller, Glenn, Michael W. Kattan, and Howard I. Scher. "Improving the decision to pursue a phase 3 clinical trial 
by adjusting for patient-specific factors in evaluating phase 2 treatment efficacy data." Medical Decision Making 
27.4 (2007): 380-386.
}
\examples{
library(MASS)

}
\author{
Daniel D Sjoberg \email{sjobergd@mskcc.org}
}
